+++ src/types.tsx
@@ -1,0 +1,3 @@
+export interface PeerError {
+  type: string;
+}